ExtView: Allow window type and flags to be overridden
KeyguardExternalViews need to have a different window type
as well as different flags set when creating the window.  this
patch allows those to be overridden when extending the
ExternalViewProviderService.  This patch also makes sure that
the width and height of the external view window is always the
same as the display size.

Change-Id: Ic5fc0d211e9fee0e610d55a3620e6e322defe231